+++ examples/basic/www/js/index.js
@@ -12,17 +12,29 @@ const app = {
   onDeviceReady() {
     this.receivedEvent('deviceready')
 
-    this.checkIsLoaded().then(() => this.showAds())
+    this.showAds()
+    this.checkIsLoaded()
   },
 
   checkIsLoaded() {
     return admob.interstitial.isLoaded().then(result => {
-      alert(result)
+      console.log(result)
       return result
     })
   },
 
   showAds() {
+    const bannerTop = new admob.BannerAd({
+      adUnitId: 'ca-app-pub-3940256099942544/6300978111',
+    })
+    bannerTop.show({ position: 'top' })
+    const banner = new admob.BannerAd({
+      adUnitId: 'ca-app-pub-3940256099942544/2934735716',
+    })
+    banner.show({ position: 'bottom' })
+  },
+
+  showAdsOld() {
     admob.banner.show({ id: 'test', npa: '1' }).catch(console.log)
     admob.interstitial
       .load({ id: 'test' })