+++ packages/cordova/src/ios/AMSBanner.swift
@@ -4,8 +4,12 @@ class AMSBanner: AMSAdBase, GADBannerViewDelegate {
     var position: String!
     var constraintsToHide: [NSLayoutConstraint]!
 
-    var view: UIView {
+    var rootView: UIView {
         return self.plugin.viewController.view
+    }
+
+    var mainView: UIView {
+        return self.plugin.webView
     }
 
     init(id: Int, adUnitID: String, adSize: GADAdSize, position: String) {
@@ -14,8 +18,8 @@ class AMSBanner: AMSAdBase, GADBannerViewDelegate {
         self.adSize = adSize
         self.position = position
         self.constraintsToHide = [
-            self.plugin.webView.topAnchor.constraint(equalTo: view.topAnchor),
-            self.plugin.webView.bottomAnchor.constraint(equalTo: view.bottomAnchor)
+            mainView.topAnchor.constraint(equalTo: rootView.topAnchor),
+            mainView.bottomAnchor.constraint(equalTo: rootView.bottomAnchor)
         ]
     }
 
@@ -49,21 +53,21 @@ class AMSBanner: AMSAdBase, GADBannerViewDelegate {
     }
 
     func addBannerViewToView(_ bannerView: UIView) {
-        self.plugin.webView.translatesAutoresizingMaskIntoConstraints = false
+        mainView.translatesAutoresizingMaskIntoConstraints = false
         bannerView.translatesAutoresizingMaskIntoConstraints = false
-        view.addSubview(bannerView)
+        rootView.addSubview(bannerView)
         if #available(iOS 11.0, *) {
             positionBannerInSafeArea(bannerView)
 
             let background = UIView()
             background.translatesAutoresizingMaskIntoConstraints = false
             background.backgroundColor = .black
-            view.addSubview(background)
-            view.sendSubviewToBack(background)
+            rootView.addSubview(background)
+            rootView.sendSubviewToBack(background)
             NSLayoutConstraint.activate([
-                background.leadingAnchor.constraint(equalTo: view.leadingAnchor),
-                background.trailingAnchor.constraint(equalTo: view.trailingAnchor),
-                background.bottomAnchor.constraint(equalTo: view.bottomAnchor),
+                background.leadingAnchor.constraint(equalTo: rootView.leadingAnchor),
+                background.trailingAnchor.constraint(equalTo: rootView.trailingAnchor),
+                background.bottomAnchor.constraint(equalTo: rootView.bottomAnchor),
                 background.topAnchor.constraint(equalTo: bannerView.topAnchor)
             ])
         } else {
@@ -73,38 +77,42 @@ class AMSBanner: AMSAdBase, GADBannerViewDelegate {
 
     @available (iOS 11, *)
     func positionBannerInSafeArea(_ bannerView: UIView) {
-        let guide: UILayoutGuide = view.safeAreaLayoutGuide
+        let guide: UILayoutGuide = rootView.safeAreaLayoutGuide
         var constraints = [
             bannerView.centerXAnchor.constraint(equalTo: guide.centerXAnchor),
-            self.plugin.webView.leadingAnchor.constraint(equalTo: view.leadingAnchor),
-            self.plugin.webView.trailingAnchor.constraint(equalTo: view.trailingAnchor)
+            bannerView.heightAnchor.constraint(equalToConstant: bannerView.frame.height),
+            mainView.widthAnchor.constraint(equalTo: rootView.widthAnchor)
         ]
         if position == "top" {
+            let c = mainView.bottomAnchor.constraint(equalTo: rootView.bottomAnchor)
+            c.priority = UILayoutPriority(999)
             constraints += [
                 bannerView.topAnchor.constraint(equalTo: guide.topAnchor),
-                self.plugin.webView.topAnchor.constraint(equalTo: bannerView.bottomAnchor),
-                self.plugin.webView.bottomAnchor.constraint(equalTo: view.bottomAnchor)
+                c,
+                mainView.topAnchor.constraint(equalTo: bannerView.bottomAnchor)
             ]
         } else {
+            let c = mainView.topAnchor.constraint(equalTo: rootView.topAnchor)
+            c.priority = UILayoutPriority(999)
             constraints += [
                 bannerView.bottomAnchor.constraint(equalTo: guide.bottomAnchor),
-                self.plugin.webView.topAnchor.constraint(equalTo: view.topAnchor),
-                self.plugin.webView.bottomAnchor.constraint(equalTo: bannerView.topAnchor)
+                c,
+                mainView.bottomAnchor.constraint(equalTo: bannerView.topAnchor)
             ]
         }
         NSLayoutConstraint.activate(constraints)
     }
 
     func positionBanner(_ bannerView: UIView) {
-        view.addConstraint(NSLayoutConstraint(item: bannerView,
+        rootView.addConstraint(NSLayoutConstraint(item: bannerView,
                                               attribute: .centerX,
                                               relatedBy: .equal,
-                                              toItem: view,
+                                              toItem: rootView,
                                               attribute: .centerX,
                                               multiplier: 1,
                                               constant: 0))
         if position == "top" {
-            view.addConstraint(NSLayoutConstraint(item: bannerView,
+            rootView.addConstraint(NSLayoutConstraint(item: bannerView,
                                                   attribute: .top,
                                                   relatedBy: .equal,
                                                   toItem: plugin.viewController.topLayoutGuide,
@@ -112,7 +120,7 @@ class AMSBanner: AMSAdBase, GADBannerViewDelegate {
                                                   multiplier: 1,
                                                   constant: 0))
         } else {
-            view.addConstraint(NSLayoutConstraint(item: bannerView,
+            rootView.addConstraint(NSLayoutConstraint(item: bannerView,
                                                   attribute: .bottom,
                                                   relatedBy: .equal,
                                                   toItem: plugin.viewController.bottomLayoutGuide,