+++ scripts/utils.ts
@@ -1,0 +1,4 @@
+import * as path from 'path'
+
+export const pkgsDirJoin = (...args: string[]) =>
+  path.join(__dirname, '../packages', ...args)