+++ website/sidebars.js
@@ -1,0 +1,10 @@
+module.exports = {
+  docs: {
+    'Getting Started': ['installation'],
+    'Display Ads': ['show-banner', 'show-interstitial', 'show-reward-video'],
+    Guides: ['test-ads', 'volume-control', 'designed-for-families'],
+    'Ionic Guides': ['ionic'],
+    'API Reference': ['ad-request-options'],
+    Troubleshooting: ['faq'],
+  },
+}